# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "E-Ticaret API",
        "description": "JWT kimlik doğrulama ile E-Ticaret uygulaması için API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/products": {
            "get": {
                "tags": [
                    "Ürünler"
                ],
                "summary": "Filtreleme ve sayfalama ile tüm ürünleri getir",
                "description": "Ürünleri kategori, fiyat aralığı ve arama ile filtreleyebilir, sayfalama yapabilirsiniz",
                "operationId": "d97d1c3727cbac49708b85f7d11f7d58",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Sayfa numarası",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Sayfa başına ürün sayısı",
                        "schema": {
                            "type": "integer",
                            "default": 20
                        }
                    },
                    {
                        "name": "category_id",
                        "in": "query",
                        "description": "Kategori ID'sine göre filtrele",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "min_price",
                        "in": "query",
                        "description": "Minimum fiyat filtresi",
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "max_price",
                        "in": "query",
                        "description": "Maksimum fiyat filtresi",
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Ürün adında arama yap",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ürünler başarıyla getirildi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "products": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                },
                                                "pagination": {
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Yetkisiz erişim - Geçersiz veya eksik JWT token"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Ürünler"
                ],
                "summary": "Yeni ürün oluştur",
                "description": "Yeni bir ürün oluşturur (Sadece Admin)",
                "operationId": "785644b7d7dd38af4b51dafd0fbc1b6d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "price",
                                    "stock_quantity",
                                    "category_id"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Ürün adı",
                                        "type": "string",
                                        "example": "iPhone 15"
                                    },
                                    "description": {
                                        "description": "Ürün açıklaması",
                                        "type": "string",
                                        "example": "En son iPhone modeli"
                                    },
                                    "price": {
                                        "description": "Ürün fiyatı",
                                        "type": "number",
                                        "example": 999.99
                                    },
                                    "stock_quantity": {
                                        "description": "Stok miktarı",
                                        "type": "integer",
                                        "example": 50
                                    },
                                    "category_id": {
                                        "description": "Kategori ID'si",
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Ürün başarıyla oluşturuldu",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Ürün başarıyla oluşturuldu"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Yetkisiz erişim - Geçersiz veya eksik JWT token"
                    },
                    "403": {
                        "description": "Yasak - Admin erişimi gerekli"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/products/{id}": {
            "get": {
                "tags": [
                    "Ürünler"
                ],
                "summary": "Belirli bir ürünü getir",
                "description": "ID'si verilen ürünün detaylarını getirir",
                "operationId": "c87fb2e4d617e7b36c712cad358acdb8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Ürün ID'si",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ürün başarıyla getirildi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Yetkisiz erişim - Geçersiz veya eksik JWT token"
                    },
                    "404": {
                        "description": "Ürün bulunamadı"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Ürünler"
                ],
                "summary": "Ürün güncelle",
                "description": "Belirli bir ürünü günceller (Sadece Admin)",
                "operationId": "416e4959a3919547eb609d87e3db9135",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Güncellenecek ürünün ID'si",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Ürün adı",
                                        "type": "string",
                                        "example": "iPhone 15 Pro"
                                    },
                                    "description": {
                                        "description": "Ürün açıklaması",
                                        "type": "string",
                                        "example": "Güncellenmiş iPhone modeli"
                                    },
                                    "price": {
                                        "description": "Ürün fiyatı",
                                        "type": "number",
                                        "example": 1099.99
                                    },
                                    "stock_quantity": {
                                        "description": "Stok miktarı",
                                        "type": "integer",
                                        "example": 75
                                    },
                                    "category_id": {
                                        "description": "Kategori ID'si",
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ürün başarıyla güncellendi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Ürün başarıyla güncellendi"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Yetkisiz erişim - Geçersiz veya eksik JWT token"
                    },
                    "404": {
                        "description": "Ürün bulunamadı"
                    },
                    "403": {
                        "description": "Yasak - Admin erişimi gerekli"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Ürünler"
                ],
                "summary": "Ürün sil",
                "description": "Belirli bir ürünü siler (Sadece Admin)",
                "operationId": "b49ce9f3cb168b233245548fb4ef36d3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Silinecek ürünün ID'si",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ürün başarıyla silindi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Ürün başarıyla silindi"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Yetkisiz erişim - Geçersiz veya eksik JWT token"
                    },
                    "404": {
                        "description": "Ürün bulunamadı"
                    },
                    "403": {
                        "description": "Yasak - Admin erişimi gerekli"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Login veya register endpoint'inden alınan JWT token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}